refactor: dynamic query params using axios
