+++ src/api/types/index.ts
@@ -1,2 +1,4 @@
 export * from './user'
 export * from './public_config'
+export * from './pagination'
+export * from './transaction'
