+++ src/api/types/pagination.ts
@@ -1,0 +1,6 @@
+export type PaginationParams = {
+  page: number
+  page_size: number // this is a backend API requirement, not a TypeScript naming convention
+  order: string
+  sort: 'desc' | 'asc'
+}
